finish basic backend capacilities
