+++ src/components/TestButton.js
@@ -6,7 +6,7 @@ export default function TestButton() {
 
   const fetchUser = async () => {
     try {
-      const response = await fetch("/api/v1/test/gemini", {
+      const response = await fetch("/api/v1/pdf/", {
         method: "GET",
         credentials: "include",
         headers: { "Content-Type": "application/json" },
@@ -17,7 +17,6 @@ export default function TestButton() {
       }
 
       const data = await response.json();
-      setUserData(data);
     } catch (err) {
       setError(err.message);
     }
